Move CreateRoutes from WebFinalComponent to WebInitialComponent
